GROW-81 fix: 자산적립 추이 api, 자산 상승률 오류 수정
