+++ app/module/asset/services/asset_service.py
@@ -1,22 +1,26 @@
+from typing import Any
+
 from sqlalchemy.ext.asyncio import AsyncSession
 
+from app.module.asset.enum import ASSETNAME, AmountUnit
 from app.module.asset.model import Asset
 from app.module.asset.repository.asset_repository import AssetRepository
 from app.module.asset.schema import AssetStockPutRequest
-from app.module.asset.enum import AmountUnit, ASSETNAME
 
 
 class AssetService:
     @staticmethod
     def get_average_investment_with_dividend_year(
-        total_invest_amount:float,
-        total_dividend_amount:float,
-        months:float
+        total_invest_amount: float, total_dividend_amount: float, months: float
     ) -> float:
         average_invest_amount_month = total_invest_amount / months if total_invest_amount > 0.0 else 0.0
         average_dividend_month = total_dividend_amount / months if total_dividend_amount > 0.0 else 0.0
-        return (average_invest_amount_month + average_dividend_month) * 12 if average_invest_amount_month + average_dividend_month > 0 else 0.0
-    
+        return (
+            (average_invest_amount_month + average_dividend_month) * 12
+            if average_invest_amount_month + average_dividend_month > 0
+            else 0.0
+        )
+
     @staticmethod
     def calculate_trend_values(
         total_asset_amount: float,
@@ -25,33 +29,30 @@ class AssetService:
         total_profit_rate_real: float,
         years: int,
     ) -> tuple[dict[str, list[float]], dict[str, list[float]], str]:
-        values1 = {"values": [], "name": ASSETNAME.ESTIMATE_ASSET}
-        values2 = {"values": [], "name": ASSETNAME.REAL_ASSET}
-        
+        values1: dict[str, Any] = {"values": [], "name": ASSETNAME.ESTIMATE_ASSET}
+        values2: dict[str, Any] = {"values": [], "name": ASSETNAME.REAL_ASSET}
+
         current_value1 = total_asset_amount
         current_value2 = total_asset_amount
 
         for _ in range(years):
             current_value1 += increase_invest_year
-            current_value1 *= 1 + total_profit_rate
+            current_value1 *= (1 + total_profit_rate / 100)
             values1["values"].append(current_value1)
 
             current_value2 += increase_invest_year
-            current_value2 *= 1 + total_profit_rate_real
+            current_value2 *= (1 + total_profit_rate_real / 100)
             values2["values"].append(current_value2)
 
         if total_asset_amount >= 100000000:
-            unit = AmountUnit.BILLION_WON
             values1["values"] = [v / 100000000 for v in values1["values"]]
             values2["values"] = [v / 100000000 for v in values2["values"]]
+            unit = AmountUnit.BILLION_WON
         else:
             unit = AmountUnit.MILLION_WON
 
         return values1, values2, unit
-    
-    
-    
-    
+
     @staticmethod
     async def get_asset_map(session: AsyncSession, asset_id: int) -> dict[int, Asset] | None:
         asset = await AssetRepository.get_asset_by_id(session, asset_id)
