+++ test/app/module/asset/test_schema.py
@@ -4,18 +4,13 @@ from redis.asyncio import Redis
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.module.asset.enum import AssetType
+from app.module.asset.facades.asset_facade import AssetFacade
+from app.module.asset.facades.dividend_facade import DividendFacade
 from app.module.asset.model import Asset
 from app.module.asset.repository.asset_repository import AssetRepository
 from app.module.asset.schema import AssetStockResponse, UpdateAssetFieldRequest
 from app.module.asset.services.asset_field_service import AssetFieldService
-from app.module.asset.services.asset_stock_service import AssetStockService
-from app.module.asset.services.dividend_service import DividendService
-from app.module.asset.services.exchange_rate_service import ExchangeRateService
-from app.module.asset.services.stock_daily_service import StockDailyService
-from app.module.asset.services.stock_service import StockService
 from app.module.auth.constant import DUMMY_USER_ID
-from app.module.asset.facades.asset_facade import AssetFacade
-from app.module.asset.facades.dividend_facade import DividendFacade
 
 
 class TestUpdateAssetFieldRequest:
@@ -103,14 +98,11 @@ class TestAssetStockResponse:
         assets: list[Asset] = await AssetRepository.get_eager(session, DUMMY_USER_ID, AssetType.STOCK)
         asset_fields = await AssetFieldService.get_asset_field(session, DUMMY_USER_ID)
         stock_assets: list[dict] = await AssetFacade.get_stock_assets(
-                session=session,
-                redis_client=redis_client,
-                assets=assets,
-                asset_fields=asset_fields
-            )
-        
+            session=session, redis_client=redis_client, assets=assets, asset_fields=asset_fields
+        )
+
         total_asset_amount = await AssetFacade.get_total_asset_amount(session, redis_client, assets)
-        total_invest_amount = await AssetFacade.get_total_investment_amount(session, redis_client, assets)   
+        total_invest_amount = await AssetFacade.get_total_investment_amount(session, redis_client, assets)
         total_dividend_amount = await DividendFacade.get_total_dividend(session, redis_client, assets)
 
         # When
